refactor: integrate tests
