--- test/test_fixed_communication.cc
@@ -1,347 +1,0 @@
-#include <gtest/gtest.h>
-#include <thread>
-#include <chrono>
-#include <memory>
-#include <atomic>
-#include <condition_variable>
-#include <mutex>
-#include <iostream>
-
-#include "unilink/unilink.hpp"
-
-using namespace unilink;
-using namespace std::chrono_literals;
-
-class FixedCommunicationTest : public ::testing::Test {
-protected:
-    void SetUp() override {
-        data_received_.clear();
-        connection_established_ = false;
-        error_occurred_ = false;
-        server_ready_ = false;
-        client_connected_ = false;
-    }
-    
-    void TearDown() override {
-        if (client_) {
-            std::cout << "Stopping client..." << std::endl;
-            client_->stop();
-        }
-        if (server_) {
-            std::cout << "Stopping server..." << std::endl;
-            server_->stop();
-        }
-        
-        // 충분한 시간을 두고 정리
-        std::this_thread::sleep_for(std::chrono::milliseconds(500));
-    }
-
-    // 테스트용 포트 번호
-    uint16_t getTestPort() {
-        static std::atomic<uint16_t> port_counter{30000};
-        return port_counter.fetch_add(1);
-    }
-
-    // 서버 준비 대기 (리스닝 상태 확인)
-    void waitForServerReady(std::chrono::milliseconds timeout = 3000ms) {
-        auto start = std::chrono::steady_clock::now();
-        while (std::chrono::steady_clock::now() - start < timeout) {
-            if (server_ready_.load()) {
-                return;
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(100));
-        }
-    }
-
-    // 클라이언트 연결 대기
-    void waitForClientConnection(std::chrono::milliseconds timeout = 3000ms) {
-        auto start = std::chrono::steady_clock::now();
-        while (std::chrono::steady_clock::now() - start < timeout) {
-            if (client_connected_.load()) {
-                return;
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(100));
-        }
-    }
-
-    // 데이터 수신 대기
-    void waitForData(std::chrono::milliseconds timeout = 2000ms) {
-        auto start = std::chrono::steady_clock::now();
-        while (std::chrono::steady_clock::now() - start < timeout) {
-            if (!data_received_.empty()) {
-                return;
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(100));
-        }
-    }
-
-protected:
-    std::shared_ptr<wrapper::TcpServer> server_;
-    std::shared_ptr<wrapper::TcpClient> client_;
-    
-    std::mutex mtx_;
-    std::condition_variable cv_;
-    std::vector<std::string> data_received_;
-    std::atomic<bool> connection_established_{false};
-    std::atomic<bool> error_occurred_{false};
-    std::atomic<bool> server_ready_{false};
-    std::atomic<bool> client_connected_{false};
-    std::string last_error_;
-};
-
-/**
- * @brief 서버 시작 및 리스닝 상태 확인 테스트
- */
-TEST_F(FixedCommunicationTest, ServerStartAndListen) {
-    // --- Setup ---
-    uint16_t test_port = getTestPort();
-    std::cout << "Testing server start with port: " << test_port << std::endl;
-    
-    // --- Test Logic ---
-    // 1. 서버 생성
-    std::cout << "Creating server..." << std::endl;
-    server_ = builder::UnifiedBuilder::tcp_server(test_port)
-        .auto_start(false)  // 수동 시작
-        .on_connect([this]() {
-            std::cout << "Server: Client connected!" << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            connection_established_ = true;
-            cv_.notify_one();
-        })
-        .on_error([this](const std::string& error) {
-            std::cout << "Server error: " << error << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            error_occurred_ = true;
-            last_error_ = error;
-            cv_.notify_one();
-        })
-        .build();
-    
-    ASSERT_NE(server_, nullptr) << "Server creation failed";
-    std::cout << "Server created successfully" << std::endl;
-    
-    // 2. 서버 시작
-    std::cout << "Starting server..." << std::endl;
-    server_->start();
-    
-    // 3. 서버가 리스닝 상태가 될 때까지 대기
-    std::cout << "Waiting for server to be ready..." << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-    
-    // 4. 서버 상태 확인
-    std::cout << "Server state after 2s:" << std::endl;
-    std::cout << "  is_connected(): " << (server_->is_connected() ? "true" : "false") << std::endl;
-    std::cout << "  Error occurred: " << (error_occurred_.load() ? "true" : "false") << std::endl;
-    if (error_occurred_.load()) {
-        std::cout << "  Last error: " << last_error_ << std::endl;
-    }
-    
-    // --- Verification ---
-    // 서버가 에러 없이 시작되었는지 확인
-    EXPECT_FALSE(error_occurred_.load()) << "Server failed to start: " << last_error_;
-    
-    // 서버가 생성되었는지 확인 (is_connected()는 클라이언트 연결 시에만 true)
-    EXPECT_TRUE(server_ != nullptr);
-}
-
-/**
- * @brief 클라이언트 연결 테스트
- */
-TEST_F(FixedCommunicationTest, ClientConnection) {
-    // --- Setup ---
-    uint16_t test_port = getTestPort();
-    std::cout << "Testing client connection with port: " << test_port << std::endl;
-    
-    // --- Test Logic ---
-    // 1. 서버 생성 및 시작
-    std::cout << "Creating and starting server..." << std::endl;
-    server_ = builder::UnifiedBuilder::tcp_server(test_port)
-        .auto_start(true)  // 자동 시작
-        .on_connect([this]() {
-            std::cout << "Server: Client connected!" << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            connection_established_ = true;
-            server_ready_ = true;
-            cv_.notify_one();
-        })
-        .on_error([this](const std::string& error) {
-            std::cout << "Server error: " << error << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            error_occurred_ = true;
-            last_error_ = error;
-            cv_.notify_one();
-        })
-        .build();
-    
-    ASSERT_NE(server_, nullptr) << "Server creation failed";
-    
-    // 서버 시작 대기
-    std::cout << "Waiting for server to start..." << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-    
-    // 2. 클라이언트 생성
-    std::cout << "Creating client..." << std::endl;
-    client_ = builder::UnifiedBuilder::tcp_client("127.0.0.1", test_port)
-        .auto_start(true)  // 자동 시작
-        .on_connect([this]() {
-            std::cout << "Client: Connected to server!" << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            client_connected_ = true;
-            cv_.notify_one();
-        })
-        .on_error([this](const std::string& error) {
-            std::cout << "Client error: " << error << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            error_occurred_ = true;
-            last_error_ = error;
-            cv_.notify_one();
-        })
-        .build();
-    
-    ASSERT_NE(client_, nullptr) << "Client creation failed";
-    
-    // 클라이언트 연결 대기
-    std::cout << "Waiting for client to connect..." << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(5000));
-    
-    // 3. 상태 확인
-    std::cout << "Final states:" << std::endl;
-    std::cout << "  Server is_connected(): " << (server_->is_connected() ? "true" : "false") << std::endl;
-    std::cout << "  Client is_connected(): " << (client_->is_connected() ? "true" : "false") << std::endl;
-    std::cout << "  Connection established: " << (connection_established_.load() ? "true" : "false") << std::endl;
-    std::cout << "  Client connected: " << (client_connected_.load() ? "true" : "false") << std::endl;
-    std::cout << "  Error occurred: " << (error_occurred_.load() ? "true" : "false") << std::endl;
-    if (error_occurred_.load()) {
-        std::cout << "  Last error: " << last_error_ << std::endl;
-    }
-    
-    // --- Verification ---
-    // 에러가 발생하지 않았는지 확인
-    EXPECT_FALSE(error_occurred_.load()) << "Connection failed: " << last_error_;
-    
-    // 서버와 클라이언트가 모두 생성되었는지 확인
-    EXPECT_TRUE(server_ != nullptr);
-    EXPECT_TRUE(client_ != nullptr);
-}
-
-/**
- * @brief 실제 데이터 통신 테스트
- */
-TEST_F(FixedCommunicationTest, RealDataCommunication) {
-    // --- Setup ---
-    uint16_t test_port = getTestPort();
-    std::cout << "Testing data communication with port: " << test_port << std::endl;
-    
-    // --- Test Logic ---
-    // 1. 서버 생성
-    std::cout << "Creating server..." << std::endl;
-    server_ = builder::UnifiedBuilder::tcp_server(test_port)
-        .auto_start(true)
-        .on_data([this](const std::string& data) {
-            std::cout << "Server received: " << data << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            data_received_.push_back("SERVER: " + data);
-            cv_.notify_one();
-        })
-        .on_connect([this]() {
-            std::cout << "Server: Client connected!" << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            connection_established_ = true;
-            server_ready_ = true;
-            cv_.notify_one();
-        })
-        .on_error([this](const std::string& error) {
-            std::cout << "Server error: " << error << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            error_occurred_ = true;
-            last_error_ = error;
-            cv_.notify_one();
-        })
-        .build();
-    
-    ASSERT_NE(server_, nullptr) << "Server creation failed";
-    
-    // 서버 시작 대기
-    std::cout << "Waiting for server to start..." << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
-    
-    // 2. 클라이언트 생성
-    std::cout << "Creating client..." << std::endl;
-    client_ = builder::UnifiedBuilder::tcp_client("127.0.0.1", test_port)
-        .auto_start(true)
-        .on_data([this](const std::string& data) {
-            std::cout << "Client received: " << data << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            data_received_.push_back("CLIENT: " + data);
-            cv_.notify_one();
-        })
-        .on_connect([this]() {
-            std::cout << "Client: Connected!" << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            client_connected_ = true;
-            cv_.notify_one();
-        })
-        .on_error([this](const std::string& error) {
-            std::cout << "Client error: " << error << std::endl;
-            std::lock_guard<std::mutex> lock(mtx_);
-            error_occurred_ = true;
-            last_error_ = error;
-            cv_.notify_one();
-        })
-        .build();
-    
-    ASSERT_NE(client_, nullptr) << "Client creation failed";
-    
-    // 클라이언트 연결 대기
-    std::cout << "Waiting for client to connect..." << std::endl;
-    std::this_thread::sleep_for(std::chrono::milliseconds(3000));
-    
-    // 3. 데이터 전송 시도
-    if (client_->is_connected()) {
-        std::cout << "Sending test message..." << std::endl;
-        const std::string test_message = "Hello from client!";
-        client_->send(test_message);
-        
-        // 데이터 수신 대기
-        std::cout << "Waiting for data reception..." << std::endl;
-        waitForData(3000ms);
-        
-        std::cout << "Data received count: " << data_received_.size() << std::endl;
-        for (const auto& data : data_received_) {
-            std::cout << "Received: " << data << std::endl;
-        }
-        
-        // --- Verification ---
-        EXPECT_FALSE(data_received_.empty()) << "No data received by server";
-        
-        // 서버가 데이터를 받았는지 확인
-        bool server_received = false;
-        for (const auto& data : data_received_) {
-            if (data.find("SERVER: " + test_message) != std::string::npos) {
-                server_received = true;
-                break;
-            }
-        }
-        EXPECT_TRUE(server_received) << "Server did not receive the message";
-        
-    } else {
-        std::cout << "Client not connected, skipping data transmission" << std::endl;
-        std::cout << "Final states:" << std::endl;
-        std::cout << "  Server is_connected(): " << (server_->is_connected() ? "true" : "false") << std::endl;
-        std::cout << "  Client is_connected(): " << (client_->is_connected() ? "true" : "false") << std::endl;
-        std::cout << "  Connection established: " << (connection_established_.load() ? "true" : "false") << std::endl;
-        std::cout << "  Client connected: " << (client_connected_.load() ? "true" : "false") << std::endl;
-        std::cout << "  Error occurred: " << (error_occurred_.load() ? "true" : "false") << std::endl;
-        if (error_occurred_.load()) {
-            std::cout << "  Last error: " << last_error_ << std::endl;
-        }
-        
-        // 연결이 실패한 경우에도 테스트는 통과 (네트워크 환경에 따라)
-        GTEST_SKIP() << "Client could not connect to server (network environment dependent)";
-    }
-}
-
-int main(int argc, char** argv) {
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
